requirements.txt
selenium
google-generativeai
pandas
python-dotenv
fastapi
uvicorn[standard]